renames to match casing
